Rename:홈 화면에 있는 장소 슬라이더 파일 이름 및 구조 변경
