+++ app/feature/place/components/PlaceDetail/PlaceImageSlider.tsx
@@ -7,8 +7,10 @@ import Image from "next/image";
 import Slider from "react-slick";
 import { useEffect, useState } from "react";
 import { PlaceDetailInfoProps } from "@feature/place/queries/dto/place-detail";
+import PlaceImageSliderCount from "./PlaceImageSliderCount";
 
-export default function PlaceDetailImageSlider({
+//Organism
+export default function PlaceImageSlider({
   imgUrlList,
 }: Pick<PlaceDetailInfoProps, "imgUrlList">) {
   const [imgIndex, setImgIndex] = useState(0);
@@ -24,9 +26,7 @@ export default function PlaceDetailImageSlider({
   }, []);
   return (
     <div className="relative w-[100%] overflow-hidden">
-      <div className="flex justify-center items-center absolute w-[3.9rem] h-[2rem] rounded-[1000px] z-10 bottom-[3.9rem] right-[2rem] bg-black opacity-[0.8] text-white body3-medium">
-        {imgIndex + 1}/2
-      </div>
+      <PlaceImageSliderCount imgIndex={imgIndex} />
       <Slider {...sliderSettings}>
         {imgUrlList.map((image, i) => (
           <div
